Project fixing and state manager improvement.
The project works fine now.
