--- Modular Character Controller.csproj.old.1
@@ -1,11 +1,0 @@
-<Project Sdk="Godot.NET.Sdk/4.0.3">
-  <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
-    <EnableDynamicLoading>true</EnableDynamicLoading>
-    <RootNamespace>ModularCharacterController</RootNamespace>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-  <ItemGroup>
-    <!-- <PackageReference Include="Firebelley.GodotUtilities" Version="4.0.2" /> -->
-  </ItemGroup>
-</Project>
