+++ Scenes/Entities/Player/CustomStates/Body/BIdle_BodyState.tscn
@@ -1,6 +1,6 @@
 [gd_scene load_steps=2 format=3 uid="uid://c8gvkvxuj2bq4"]
 
-[ext_resource type="Script" path="res://Scripts/Entities/Player/CustomStates/Body/BIdle_BodyState.cs" id="1_30hei"]
+[ext_resource type="Script" path="res://Scripts/Entities/Player/State/CustomStates/Body/BIdle_BodyState.cs" id="1_0w7q2"]
 
 [node name="BIdle_BodyState" type="Node"]
-script = ExtResource("1_30hei")
+script = ExtResource("1_0w7q2")
