+++ Scenes/Entities/Player/CustomStates/Body/BNormal_BodyState.tscn
@@ -1,11 +1,11 @@
 [gd_scene load_steps=4 format=3 uid="uid://h18a1s5d4ygf"]
 
-[ext_resource type="Script" path="res://Scripts/Entities/Player/CustomStates/Body/BNormal_BodyState.cs" id="1_gyfy0"]
+[ext_resource type="Script" path="res://Scripts/Entities/Player/State/CustomStates/Body/BNormal_BodyState.cs" id="1_y2cpf"]
 [ext_resource type="PackedScene" path="res://Archive/MovementComponents/VelocityComponent.tscn" id="2_emg0u"]
 [ext_resource type="PackedScene" uid="uid://bwl5a2ndhny5a" path="res://Archive/MovementComponents/PlayerMove_KeyComponent.tscn" id="3_v3c1h"]
 
 [node name="BNormal_BodyState" type="Node" groups=["BodyStates"]]
-script = ExtResource("1_gyfy0")
+script = ExtResource("1_y2cpf")
 
 [node name="VelocityComponent" parent="." instance=ExtResource("2_emg0u")]
