+++ Scripts/Components/StateMachine/StateMachine.cs
@@ -2,7 +2,7 @@ using Godot;
 
 // This only works for CharacterBody2D and a separate state machine is neccasary for other bodies.
 
-[Tool]
+
 public partial class StateMachine : Node
 {
     [Export]
@@ -61,29 +61,5 @@ public partial class StateMachine : Node
     public void PhysicsProcess()
     {
         CurrentState.PhysicsProcess();
-    }
-
-    // Checks if there are child nodes and if any child node is not in BodyState group.
-    public override string[] _GetConfigurationWarnings()
-    {
-        string[] empty = new string[0];
-        string[] warnings = new string[1];
-
-        if (GetChildCount() == 0)
-        {
-            warnings[0] = "Assigns state nodes.";
-            return warnings;
-        }
-
-        foreach (Node node in GetChildren())
-        {
-            if (node.IsInGroup("BodyStates") == false)
-            {
-                warnings[0] = "Child node is not body state.";
-                return warnings;
-            }
-        }
-
-        return empty;
     }
 }
